idlc: implemented 'module', 'struct', 'union', and 'enum'
The parsed 'module', 'struct', 'union' and 'enum' are processed
and added to the type declarations.

Tests are implemented with a stringifier which builds a sting
representation from the type declarations. There are actually
two implementations of this (while testing some 'abstract'
api's). One implementation is based on iterators that traverse
over the type declarations (see type_api.h/c). The other
implementation constructs an 'expression' that is executed
against the type declarations (see walker.h/c).
(Note that these are too are experimental and likely will be
removed in the futute.)