idlc: restructuring type tree.
This is an intermediate commit, primarily for the visitor_api.md document,
which describes ideas for using a visitor pattern the IDL back-end API.

Furthermore the new file dds_tt.h gives a restructuring of the type tree,
which will be filled during parsing and made available to the back-ends.
This is not the final version of the file.

The remainder of the code has been adapted accordingly, but except from
compiling, is not really functional. Some parts will be removed in the
future.