IDL compiler: initial commit
This includes the relevant parts of the IDL 4.2 grammar in the
files idl.l and idl.y.

The file dds_tt.h gives the structures for the type tree.

The files tt_create.h/c give a partial implementation of how
the type tree is filled from the parser. It only convers the
base types (including sequence, string, wstring, fixed and
map), module, struct, enum, and union (partially). (The processing
of const literals/values is not complete.)

The files tt_walker.h/c implement an expression based 'visitor'
pattern that is used in stringify.c to make a string of the
type tree for testing purposes, which are implemented in
tests/parser.c.

Signed-off-by: Frans Faase <frans.faase@adlinktech.com>