IDL compiler: C99 back-end for header file
In order to develop the back-end API, work was started on generating
the header file of the C99 back-end. This implementation improves on
the existing implementation with respect to recursive types.
Because the generation of the header file is rather limited with
respect to different ways to traverse the type tree, also code was
developed to generate the meta data for the source file of the C99
back-end. This was used as input for developping an API that covers
both the generation of the header file and the meta data.

The grammar was modified to include embedded structs within structs.
Support for the #pragma keylist has not been implemented. Code has
been added to determine if a struct is used within another struct.
(This is used in the source file generation to generate only code
for structs that are not part of another struct.)

Code was added to link forward struct declarations to struct
definitions. This also required matching reopening of modules by
adding a previous pointer to the dds_ts_module_t type.

A bug with processing array declarations in the grammar was fixed.

Tests where added to check the generation of the header file. There
are no tests with respect to the meta data that is generated. (The
code probably does not produce valid XML in case multiple modules
and/or recursive types are used.)

Signed-off-by: Frans Faase <frans.faase@adlinktech.com>