IDL-compiler: C99 back-end
This commit contains an implementation of the C99 back-end limited
to the grammar constructs implemented in the front-end together with
tests.

Some modifications have been made in the ddsts. In typetree.h the
type ddsts_call_path_t has been added. This is not part of the
TypeTree but needed in utility functions that depend on the path
through which a type is reached. An example of such a function, is
the function ddsts_declaration_is_key, that can be used to find out
if a member declaration of an (embedded) struct is defined as a key.
The file type_walk.h with the function ddsts_walk has been added.

The implementation of the back-end makes use of the type gen_ostream
as a generic kind of output stream with function pointers for
opening and closing 'files'. In the tests this is used to capture
the generated output into a buffer. It could, for example, also be
used to place the generated output in a specified target directory.
The file gen_c99.h declares two functions for generating the C99
back-end given the name of the input file and a pointer to a type.

In the front-end a small problem with freeing has been solved in
case of an error. For this the semantics of ddsts_free_type have
been changed, such that it returns an error when a pointer to a part
of a type is passed. A check for using a forward struct (with no
struct definition) in a member declaration has been added.

Signed-off-by: Frans Faase <faase009@planet.nl>