IDL-Compiler: adding unions to front-end
Added support for unions to the grammar and the front-end.
Due to limited support for values, only unsigned integers,
the switch types are limited. Range checks have not been
implemented yet. (Probably better to implement range checking
when range checking for expressions is implemented.)
Tests of the front-end have been added.

Signed-off-by: Frans Faase <frans.faase@adlinktech.com>